+++ src/android/com/voxeet/toolkit/VoxeetCordova.java
@@ -34,10 +34,7 @@ import com.voxeet.sdk.push.center.management.NotificationMode;
 import com.voxeet.sdk.push.center.management.VersionFilter;
 import com.voxeet.sdk.push.center.subscription.register.SubscribeInvitation;
 import com.voxeet.sdk.services.AudioService;
-import com.voxeet.sdk.services.CommandService;
 import com.voxeet.sdk.services.ConferenceService;
-import com.voxeet.sdk.services.MediaDeviceService;
-import com.voxeet.sdk.services.RecordingService;
 import com.voxeet.sdk.services.SessionService;
 import com.voxeet.sdk.services.TelemetryService;
 import com.voxeet.sdk.services.telemetry.SdkEnvironment;
@@ -86,8 +83,6 @@ public class VoxeetCordova extends CordovaPlugin {
     private static final String VERSION = "1.3.23";
 
 
-
-
     private static final String VOXEET_CORDOVA_CONSUMER_KEY = "VOXEET_CORDOVA_CONSUMER_KEY";
     private static final String VOXEET_CORDOVA_CONSUMER_SECRET = "VOXEET_CORDOVA_CONSUMER_SECRET";
     private static final String CONSUMER_KEY = "voxeet_consumer_key";
@@ -112,6 +107,7 @@ public class VoxeetCordova extends CordovaPlugin {
     private ReentrantLock lockAwaitingToken = new ReentrantLock();
     private List<TokenCallback> mAwaitingTokenCallback;
     private CallbackContext refreshAccessTokenCallbackInstance;
+    private CallbackContext onConferenceStatusUpdatedEventCallback;
     private MicrophonePermissionWait waitMicrophonePermission;
     private CordovaWebView mWebView;
 
@@ -141,7 +137,7 @@ public class VoxeetCordova extends CordovaPlugin {
         mWebView = webView;
 
         tryInitialize(cordova.getContext(), cordova.getActivity());
-        if (null != VoxeetSDK.instance()) VoxeetSDK.instance().register(this);
+        VoxeetSDK.instance().register(this);
     }
 
     public static void tryInitialize(@NonNull Context context, @NonNull Activity activity) {
@@ -167,7 +163,7 @@ public class VoxeetCordova extends CordovaPlugin {
             consumerSecretManifest = map.get(VOXEET_CORDOVA_CONSUMER_SECRET);
 
         Log.d("CORDOVA", "initialize: " + consumerKeyManifest + " " + consumerSecretManifest);
-        if (!isEmpty(consumerKeyManifest) && !isEmpty(consumerSecretManifest) && null == VoxeetSDK.instance()) {
+        if (!isEmpty(consumerKeyManifest) && !isEmpty(consumerSecretManifest) && !VoxeetSDK.instance().isInitialized()) {
             VoxeetSDK.initialize(consumerKeyManifest, consumerSecretManifest);
             internalInitialize(null, activity);
         }
@@ -177,7 +173,7 @@ public class VoxeetCordova extends CordovaPlugin {
     public void onResume(boolean multitasking) {
         super.onResume(multitasking);
 
-        if (null != VoxeetSDK.instance()) VoxeetSDK.instance().register(this);
+        VoxeetSDK.instance().register(this);
 
         //check for permission result
         //actually not testing it in the permission callback to prevent issue with flow
@@ -191,8 +187,7 @@ public class VoxeetCordova extends CordovaPlugin {
             }
         }
 
-        ConferenceService service = VoxeetSDK.conference();
-        if (null != service && service.isLive()) {
+        if (VoxeetSDK.conference().isLive()) {
             setVolumeVoiceCall();
         }
         checkForAwaitingConference(null);
@@ -203,7 +198,7 @@ public class VoxeetCordova extends CordovaPlugin {
         super.onPause(multitasking);
 
         ConferenceService service = VoxeetSDK.conference();
-        if (null != AudioService.getSoundManager() && null != service && service.isLive()) {
+        if (null != AudioService.getSoundManager() && service.isLive()) {
             AudioService.getSoundManager().requestAudioFocus();
         }
     }
@@ -236,12 +231,12 @@ public class VoxeetCordova extends CordovaPlugin {
             }
             case PermissionRefusedEvent.RESULT_CAMERA: {
                 ConferenceService service = VoxeetSDK.conference();
-                if (null != service && service.isLive()) {
+                if (service.isLive()) {
                     service.startVideo()
                             .then((result, solver) -> {
 
                             })
-                            .error(error -> error.printStackTrace());
+                            .error(Throwable::printStackTrace);
                 }
                 return;
             }
@@ -281,17 +276,22 @@ public class VoxeetCordova extends CordovaPlugin {
                     refreshAccessTokenCallback(callbackContext);
                     break;
                 case "connect":
-                    JSONObject userInfo = null;
-                    if (!args.isNull(0)) userInfo = args.getJSONObject(0);
-                    ParticipantInfo user = null;
+                    try {
+                        JSONObject userInfo = null;
+                        if (!args.isNull(0)) userInfo = args.getJSONObject(0);
+                        ParticipantInfo user = null;
 
-                    if (null != userInfo) {
-                        user = new ParticipantInfo(userInfo.getString("name"),
-                                userInfo.getString("externalId"),
-                                userInfo.getString("avatarUrl"));
+                        if (null != userInfo) {
+                            user = new ParticipantInfo(string(userInfo, "name"),
+                                    string(userInfo, "externalId"),
+                                    string(userInfo, "avatarUrl"));
+                        }
+
+                        openSession(user, callbackContext);
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                        callbackContext.error(e.getMessage());
                     }
-
-                    openSession(user, callbackContext);
                     break;
                 case "disconnect":
                     closeSession(callbackContext);
@@ -311,7 +311,6 @@ public class VoxeetCordova extends CordovaPlugin {
                 case "create":
                     try {
                         JSONObject parameters = args.getJSONObject(0);
-
 
                         String confAlias = parameters.getString("alias");
                         JSONObject object = null;
@@ -464,6 +463,12 @@ public class VoxeetCordova extends CordovaPlugin {
                 case "stopRecording":
                     stopRecording(callbackContext);
                     break;
+                case "minimize":
+                    minimize(callbackContext);
+                    break;
+                case "maximize":
+                    maximize(callbackContext);
+                    break;
                 case "setAudio3DEnabled":
                     setAudio3DEnabled(args.getBoolean(0));
                     callbackContext.success();
@@ -490,6 +495,9 @@ public class VoxeetCordova extends CordovaPlugin {
                 case "sendBroadcastMessage":
                     sendBroadcastMessage(args.getString(0), callbackContext);
                     break;
+                case "onConferenceStatusUpdatedEvent":
+                    onConferenceStatusUpdatedEventCallback = callbackContext;
+                    break;
                 default:
                     return false;
             }
@@ -510,27 +518,27 @@ public class VoxeetCordova extends CordovaPlugin {
         Users cusers = configuration.Users;
         if (null != actionBar) {
             if (actionBar.has("displayMute"))
-                cactionBar.displayMute = bool(actionBar, "displayMute");
+                cactionBar.displayMute = bool(object, "displayMute");
             if (actionBar.has("displaySpeaker"))
-                cactionBar.displaySpeaker = bool(actionBar, "displaySpeaker");
+                cactionBar.displaySpeaker = bool(object, "displaySpeaker");
             if (actionBar.has("displayCamera"))
-                cactionBar.displayCamera = bool(actionBar, "displayCamera");
+                cactionBar.displayCamera = bool(object, "displayCamera");
             if (actionBar.has("displayScreenShare"))
-                cactionBar.displayScreenShare = bool(actionBar, "displayScreenShare");
+                cactionBar.displayScreenShare = bool(object, "displayScreenShare");
             if (actionBar.has("displayLeave"))
-                cactionBar.displayLeave = bool(actionBar, "displayLeave");
+                cactionBar.displayLeave = bool(object, "displayLeave");
         }
         if (null != overlay) {
-            if (overlay.has("backgroundMaximizedColor"))
-                coverlay.background_maximized_color = integer(overlay, "backgroundMaximizedColor");
-            if (overlay.has("backgroundMinimizedColor"))
-                coverlay.background_minimized_color = integer(overlay, "backgroundMinimizedColor");
+            if (object.has("backgroundMaximizedColor"))
+                coverlay.background_maximized_color = integer(object, "backgroundMaximizedColor");
+            if (actionBar.has("backgroundMinimizedColor"))
+                coverlay.background_minimized_color = integer(object, "backgroundMinimizedColor");
         }
         if (null != users) {
-            if (users.has("speakingUserColor"))
-                cusers.speaking_user_color = integer(users, "speakingUserColor");
-            if (users.has("selectedUserColor"))
-                cusers.selected_user_color = integer(users, "selectedUserColor");
+            if (object.has("speakingUserColor"))
+                cusers.speaking_user_color = integer(object, "speakingUserColor");
+            if (actionBar.has("selectedUserColor"))
+                cusers.selected_user_color = integer(object, "selectedUserColor");
         }
     }
 
@@ -554,6 +562,16 @@ public class VoxeetCordova extends CordovaPlugin {
         }
     }
 
+    private String string(@Nullable JSONObject object, @NonNull String key) {
+        if (null == object) return null;
+        try {
+            return object.optString(key, null);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
+
     @Nullable
     private JSONObject jsonObject(@Nullable JSONObject object, @NonNull String key) {
         if (null == object) return null;
@@ -565,7 +583,6 @@ public class VoxeetCordova extends CordovaPlugin {
     }
 
     @Nullable
-
     private void defaultVideo(boolean startVideo) {
         startVideoOnJoin = startVideo;
         VoxeetPreferences.setDefaultVideoOn(startVideo);
@@ -575,9 +592,7 @@ public class VoxeetCordova extends CordovaPlugin {
     private void initialize(final String accessToken,
                             final CallbackContext callbackContext) {
         HANDLER.post(() -> {
-            Application application = (Application) cordova.getActivity().getApplicationContext();
-
-            if (null == VoxeetSDK.instance()) {
+            if (!VoxeetSDK.instance().isInitialized()) {
                 VoxeetSDK.initialize(accessToken,
                         callback -> {
                             lock(lockAwaitingToken);
@@ -588,9 +603,8 @@ public class VoxeetCordova extends CordovaPlugin {
                             postRefreshAccessToken();
                         });
 
+                VoxeetSDK.instance().register(VoxeetCordova.this);
                 internalInitialize(callbackContext, cordova.getActivity());
-                if (null != VoxeetSDK.instance())
-                    VoxeetSDK.instance().register(VoxeetCordova.this);
             } else {
                 VoxeetSDK.instance().register(VoxeetCordova.this);
                 callbackContext.success();
@@ -602,14 +616,11 @@ public class VoxeetCordova extends CordovaPlugin {
                             final String consumerSecret,
                             final CallbackContext callbackContext) {
         HANDLER.post(() -> {
-            Application application = (Application) cordova.getActivity().getApplicationContext();
-
-            if (null == VoxeetSDK.instance()) {
+            if (!VoxeetSDK.instance().isInitialized()) {
                 VoxeetSDK.initialize(consumerKey, consumerSecret);
 
+                VoxeetSDK.instance().register(VoxeetCordova.this);
                 internalInitialize(callbackContext, cordova.getActivity());
-                if (null != VoxeetSDK.instance())
-                    VoxeetSDK.instance().register(VoxeetCordova.this);
             } else {
                 VoxeetSDK.instance().register(VoxeetCordova.this);
                 callbackContext.success();
@@ -618,15 +629,13 @@ public class VoxeetCordova extends CordovaPlugin {
     }
 
     private static void internalInitialize(@Nullable final CallbackContext callbackContext, @NonNull Activity activity) {
-        ConferenceService service = VoxeetSDK.conference();
-        if (null != service)
-            service.ConferenceConfigurations.TelecomWaitingForParticipantTimeout = -1; //no timeout by default in the cordova impl
+        VoxeetSDK.conference().ConferenceConfigurations.TelecomWaitingForParticipantTimeout = -1; //no timeout by default in the cordova impl
 
         VoxeetCordova.initNotificationCenter();
 
         try {
             VoxeetSDK.notification().subscribe(new SubscribeInvitation()).execute();
-        }catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
 
@@ -650,9 +659,7 @@ public class VoxeetCordova extends CordovaPlugin {
     private void openSession(final ParticipantInfo userInfo,
                              final CallbackContext cb) {
         HANDLER.post(() -> {
-            SessionService service = VoxeetSDK.session();
-
-            if (null == service) {
+            if (!VoxeetSDK.instance().isInitialized()) {
                 cb.error(ERROR_SDK_NOT_INITIALIZED);
                 return;
             }
@@ -670,7 +677,7 @@ public class VoxeetCordova extends CordovaPlugin {
             } else {
                 _current_user = userInfo;
                 //we have an user
-                service.close()
+                VoxeetSDK.session().close()
                         .then((result, solver) -> logSelectedUser())
                         .error(error -> logSelectedUser());
             }
@@ -682,7 +689,7 @@ public class VoxeetCordova extends CordovaPlugin {
      */
     public void logSelectedUser() {
         SessionService service = VoxeetSDK.session();
-        if (null != service) {
+        if (VoxeetSDK.instance().isInitialized()) {
             service.open(_current_user)
                     .then((result, solver) -> {
                         //TODO possibility here to add management for user to be socket managed here
@@ -699,48 +706,34 @@ public class VoxeetCordova extends CordovaPlugin {
     }
 
     private void closeSession(final CallbackContext cb) {
-        SessionService service = VoxeetSDK.session();
-        if (null == service) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-
-        HANDLER.post(new Runnable() {
-            @Override
-            public void run() {
-                service.close()
-                        .then((aBoolean, solver) -> {
-                            _current_user = null;
-                            cb.success();
-                        })
-                        .error(throwable -> {
-                            _current_user = null;
-                            cb.error("Error while logging out with the server");
-                        });
-            }
-        });
+        HANDLER.post(() -> VoxeetSDK.session().close()
+                .then((aBoolean, solver) -> {
+                    _current_user = null;
+                    cb.success();
+                })
+                .error(throwable -> {
+                    _current_user = null;
+                    cb.error("Error while logging out with the server: " + throwable.getMessage());
+                }));
     }
 
     private void isUserLoggedIn(final CallbackContext cb) {
-        SessionService service = VoxeetSDK.session();
         HANDLER.post(() -> {
-            boolean logged_in = null != service && service.isSocketOpen();
+            boolean logged_in = VoxeetSDK.session().isSocketOpen();
             cb.sendPluginResult(new PluginResult(PluginResult.Status.OK, logged_in));
         });
     }
 
     private void isAudio3DEnabled(final CallbackContext cb) {
-        MediaDeviceService service = VoxeetSDK.mediaDevice();
         HANDLER.post(() -> {
-            boolean enabled = null != service && service.isAudio3DEnabled();
+            boolean enabled = VoxeetSDK.mediaDevice().isAudio3DEnabled();
             cb.sendPluginResult(new PluginResult(PluginResult.Status.OK, enabled));
         });
     }
 
     private void isTelecomMode(final CallbackContext cb) {
         HANDLER.post(() -> {
-            ConferenceService service = VoxeetSDK.conference();
-            boolean enabled = null != service && service.ConferenceConfigurations.telecomMode;
+            boolean enabled = VoxeetSDK.conference().ConferenceConfigurations.telecomMode;
 
             cb.sendPluginResult(new PluginResult(PluginResult.Status.OK, enabled));
         });
@@ -749,8 +742,7 @@ public class VoxeetCordova extends CordovaPlugin {
     private void checkForAwaitingConference(@Nullable final CallbackContext cb) {
         HANDLER.post(() -> {
             lock();
-            SessionService service = VoxeetSDK.session();
-            if (null == service) {
+            if (!VoxeetSDK.instance().isInitialized()) {
                 if (null != cb) cb.error(ERROR_SDK_NOT_INITIALIZED);
             } else {
                 CordovaIncomingBundleChecker checker = CordovaIncomingCallActivity.CORDOVA_ROOT_BUNDLE;
@@ -766,7 +758,7 @@ public class VoxeetCordova extends CordovaPlugin {
                 } else if (null != CordovaIncomingCallActivity.CORDOVA_AWAITING_BUNDLE_TO_BE_MANAGE_FOR_DECLINE) {
                     CordovaIncomingCallActivity.CORDOVA_AWAITING_BUNDLE_TO_BE_MANAGE_FOR_DECLINE.onDecline();
                 } else if (null != checker && checker.isBundleValid()) {
-                    if (null != service && service.isSocketOpen()) {
+                    if (VoxeetSDK.session().isSocketOpen()) {
                         checker.onAccept();
                         CordovaIncomingCallActivity.CORDOVA_ROOT_BUNDLE = null;
                         if (null != cb) cb.success();
@@ -786,29 +778,16 @@ public class VoxeetCordova extends CordovaPlugin {
     private void invite(final String conferenceId,
                         final List<ParticipantInfo> participants,
                         final CallbackContext cb) {
-        ConferenceService service = VoxeetSDK.conference();
-        if (null == service) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-
-        HANDLER.post(() -> service.invite(conferenceId, participants)
+        HANDLER.post(() -> VoxeetSDK.conference().invite(conferenceId, participants)
                 .then((result, solver) -> cb.success())
-                .error(throwable -> cb.error("Error while initializing the conference")));
+                .error(throwable -> cb.error("Error while initializing the conference: " + throwable.getMessage())));
     }
 
     private void create(String conferenceAlias,
                         MetadataHolder holder,
                         ParamsHolder pholder, final CallbackContext cb) {
-        ConferenceService service = VoxeetSDK.conference();
-        if (null == service) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-
-        service.create(conferenceAlias, holder, pholder)
+        VoxeetSDK.conference().create(conferenceAlias, holder, pholder)
                 .then((result) -> {
-                    //TODO add isNew
                     JSONObject object = new JSONObject();
                     try {
                         object.put("conferenceId", result.conferenceId);
@@ -819,21 +798,15 @@ public class VoxeetCordova extends CordovaPlugin {
                     }
                     cb.success(object);
                 })
-                .error(error -> cb.error("Error while creating the conference " + conferenceAlias));
+                .error(error -> cb.error("Error while creating the conference " + conferenceAlias + ": " + error.getMessage()));
     }
 
     private void broadcast(@NonNull String conferenceId, @NonNull final CallbackContext cb) {
-        ConferenceService service = VoxeetSDK.conference();
         Context context = mWebView.getContext();
         Log.d(TAG, "broadcast: broadcasting conference");
 
-        if (null == service) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-
         if (null != context && Validate.hasMicrophonePermissions(mWebView.getContext())) {
-            service.broadcast(conferenceId)
+            VoxeetSDK.conference().broadcast(conferenceId)
                     .then((result) -> {
 
                         cleanBundles();
@@ -844,7 +817,7 @@ public class VoxeetCordova extends CordovaPlugin {
 
                         cb.success();
                     })
-                    .error(error -> cb.error("Error while joining the conference " + conferenceId));
+                    .error(error -> cb.error("Error while joining the conference " + conferenceId + ": " + error.getMessage()));
         } else {
             waitMicrophonePermission = new MicrophonePermissionWait(conferenceId, cb);
             requestMicrophonePermission();
@@ -852,17 +825,11 @@ public class VoxeetCordova extends CordovaPlugin {
     }
 
     private void join(@NonNull String conferenceId, @NonNull final CallbackContext cb) {
-        ConferenceService service = VoxeetSDK.conference();
         Context context = mWebView.getContext();
         Log.d(TAG, "join: joining conference");
 
-        if (null == service) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-
         if (null != context && Validate.hasMicrophonePermissions(mWebView.getContext())) {
-            service.join(conferenceId)
+            VoxeetSDK.conference().join(conferenceId)
                     .then((result) -> {
 
                         cleanBundles();
@@ -873,7 +840,7 @@ public class VoxeetCordova extends CordovaPlugin {
 
                         cb.success();
                     })
-                    .error(error -> cb.error("Error while joining the conference " + conferenceId));
+                    .error(error -> cb.error("Error while joining the conference " + conferenceId + ": " + error.getMessage()));
         } else {
             waitMicrophonePermission = new MicrophonePermissionWait(conferenceId, cb);
             requestMicrophonePermission();
@@ -881,116 +848,72 @@ public class VoxeetCordova extends CordovaPlugin {
     }
 
     private void listen(@NonNull String conferenceId, @NonNull final CallbackContext cb) {
-        ConferenceService service = VoxeetSDK.conference();
-        Context context = mWebView.getContext();
-
-        if (null == service) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-
-        service.listen(conferenceId)
+        VoxeetSDK.conference().listen(conferenceId)
                 .then(result -> {
                     cleanBundles();
                     cb.success();
                 })
-                .error(error -> cb.error("Error while joining the conference " + conferenceId));
+                .error(error -> cb.error("Error while joining the conference " + conferenceId + ": " + error.getMessage()));
     }
 
     private void startVideo(final CallbackContext cb) {
-        ConferenceService service = VoxeetSDK.conference();
-        if (null == service) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-        service.startVideo()
+        VoxeetSDK.conference().startVideo()
                 .then((result) -> {
                     if (null != cb) cb.success();
                 })
                 .error(error -> {
                     if (null != cb) {
-                        cb.error("Error while starting video");
+                        cb.error("Error while starting video: " + error.getMessage());
                     }
                 });
     }
 
     private void stopConference(final CallbackContext cb) {
-        ConferenceService service = VoxeetSDK.conference();
-        if (null == service) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-
-        HANDLER.post(() -> service.leave()
+        HANDLER.post(() -> VoxeetSDK.conference().leave()
                 .then((ThenVoid<Boolean>) (bool) -> cb.success())
-                .error((ErrorPromise) throwable -> {
-                    cb.error("Error while leaving");
-                }));
+                .error((ErrorPromise) throwable -> cb.error("Error while leaving: " + throwable.getMessage())));
     }
 
     private void startRecording(final CallbackContext cb) {
-        RecordingService service = VoxeetSDK.recording();
-        if (null == service) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-
-        HANDLER.post(() -> service.start()
+        HANDLER.post(() -> VoxeetSDK.recording().start()
                 .then((ThenVoid<Boolean>) (bool) -> cb.success())
-                .error(throwable -> cb.error("Error while start recording")));
+                .error(throwable -> cb.error("Error while start recording: " + throwable.getMessage())));
     }
 
     private void stopRecording(final CallbackContext cb) {
-        RecordingService service = VoxeetSDK.recording();
-        if (null == service) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-
-        HANDLER.post(() -> service.stop()
+        HANDLER.post(() -> VoxeetSDK.recording().stop()
                 .then((ThenVoid<Boolean>) (bool) -> cb.success())
-                .error(new ErrorPromise() {
-                    @Override
-                    public void onError(@NonNull Throwable throwable) {
-                        cb.error("Error while stop recording");
-                    }
-                }));
+                .error(throwable -> cb.error("Error while stop recording: " + throwable.getMessage())));
+    }
+
+    private void minimize(final CallbackContext cb) {
+        VoxeetToolkit.instance().getConferenceToolkit().minimize();
+        if (cb != null) cb.success();
+    }
+
+    private void maximize(final CallbackContext cb) {
+        VoxeetToolkit.instance().getConferenceToolkit().maximize();
+        if (cb != null) cb.success();
     }
 
     private void sendBroadcastMessage(final String message, final CallbackContext cb) {
-        CommandService commandService = VoxeetSDK.command();
-        ConferenceService conferenceService = VoxeetSDK.conference();
-        if (null == commandService || null == conferenceService) {
-            cb.error(ERROR_SDK_NOT_INITIALIZED);
-            return;
-        }
-
-        String conferenceId = conferenceService.getConferenceId();
+        String conferenceId = VoxeetSDK.conference().getConferenceId();
         if (null == conferenceId || TextUtils.isEmpty(conferenceId)) {
             cb.error("You're not in a conference");
             return;
         }
 
-        HANDLER.post(new Runnable() {
-            @Override
-            public void run() {
-                commandService.send(conferenceId, message)
-                        .then((ThenVoid<Boolean>) (bool) -> cb.success())
-                        .error(throwable -> cb.error("Error while sending the message to the server"));
-            }
-        });
+        HANDLER.post(() -> VoxeetSDK.command().send(conferenceId, message)
+                .then((ThenVoid<Boolean>) (bool) -> cb.success())
+                .error(throwable -> cb.error("Error while sending the message to the server: " + throwable.getMessage())));
     }
 
     private void setAudio3DEnabled(boolean enabled) {
-        MediaDeviceService service = VoxeetSDK.mediaDevice();
-        if (null != service) service.setAudio3DEnabled(enabled);
+        VoxeetSDK.mediaDevice().setAudio3DEnabled(enabled);
     }
 
     private void setTelecomMode(boolean telecomMode) {
-        ConferenceService service = VoxeetSDK.conference();
-        if (null != service) {
-            service.ConferenceConfigurations.telecomMode = telecomMode;
-        }
+        VoxeetSDK.conference().ConferenceConfigurations.telecomMode = telecomMode;
     }
 
     private void appearMaximized(final Boolean enabled) {
@@ -1005,18 +928,15 @@ public class VoxeetCordova extends CordovaPlugin {
     }
 
     private static void defaultBuiltInSpeaker(final boolean enabled) {
-        AudioService service = VoxeetSDK.audio();
-        if (null == service) {
-            return;
-        }
-
         HANDLER.post(() -> {
             //comment lines introducing a switch in the mode
             //AudioRoute route = AudioRoute.ROUTE_PHONE;
             //if (enabled) route = AudioRoute.ROUTE_SPEAKER;
 
             VoxeetPreferences.setDefaultBuiltInSpeakerOn(enabled);
-            //service.setAudioRoute(route);
+
+            //set the contextual information which will be used by the UXKit when he join event will be fired
+            VoxeetToolkit.instance().getConferenceToolkit().Configuration.Contextual.default_speaker_on = enabled;
         });
     }
 
@@ -1120,7 +1040,7 @@ public class VoxeetCordova extends CordovaPlugin {
 
     private boolean isConnected() {
         SessionService service = VoxeetSDK.session();
-        return null != service && service.isSocketOpen();
+        return service.isSocketOpen();
     }
 
     private boolean isSameUser(@NonNull ParticipantInfo userInfo) {
@@ -1138,7 +1058,7 @@ public class VoxeetCordova extends CordovaPlugin {
     public static boolean checkForIncomingConference(@Nullable CordovaIncomingBundleChecker checker) {
         Log.d(TAG, "checkForIncomingConference: checker := " + checker);
         SessionService service = VoxeetSDK.session();
-        if (null != service && null != checker && checker.isBundleValid()) {
+        if (null != checker && checker.isBundleValid()) {
             ParticipantInfo userInfo = VoxeetPreferences.getSavedUserInfo();
 
             Log.d(TAG, "checkForIncomingConference: socket opened := " + service.isSocketOpen());
@@ -1248,6 +1168,23 @@ public class VoxeetCordova extends CordovaPlugin {
             default:
         }
         setVolumeMusic();
+
+        if (onConferenceStatusUpdatedEventCallback != null) {
+            try {
+                JSONObject jObject = new JSONObject()
+                        .put("state", event.state.toString())
+                        .put("conferenceAlias", event.conferenceAlias);
+                if (event.conference != null) {
+                    jObject.put("conferenceId", event.conference.getId());
+                }
+
+                PluginResult pluginResult = new PluginResult(PluginResult.Status.OK, jObject);
+                pluginResult.setKeepCallback(true);
+                onConferenceStatusUpdatedEventCallback.sendPluginResult(pluginResult);
+            } catch (Exception ex) {
+                Log.e(TAG, "onEvent: ConferenceJoinedSuccessEvent", ex);
+            }
+        }
     }
 
     private void setVolumeVoiceCall() {