+++ www/types/index.d.ts
@@ -1,0 +1,4 @@
+export { default as UserInfo } from "./UserInfo";
+export { default as Configuration } from "./Configurations";
+export { CreateOptions, CreateResult } from "./CreateConference";
+export { JoinOptions } from "./JoinConference";